changed size_t size

+++ include/tachyon.platform/x86/x86_common.h
@@ -26,13 +26,13 @@ typedef signed      char        int8_t;
 typedef signed      short       int16_t;
 typedef signed      int         int32_t;
 typedef signed      long long   int64_t;
-typedef signed      long        ssize_t;
+typedef intptr_t                ssize_t;
 
 typedef unsigned    char        uint8_t;
 typedef unsigned    short       uint16_t;
 typedef unsigned    int         uint32_t;
 typedef unsigned    long long   uint64_t;
-typedef unsigned    long        size_t;
+typedef uintptr_t               size_t;
 
 typedef intptr_t    intmax_t;
 typedef uintptr_t   uintmax_t;
